refactor exif read/write
